created config_util.py
