+++ mpd_subsonic_scrobbler/scrobbler_util.py
@@ -1,0 +1,15 @@
+from config_key import ConfigKey
+from subsonic_server_config import SubsonicServerConfig
+from config_util import get_indexed_env_value
+
+def get_subsonic_server_config_list() -> list[SubsonicServerConfig]:
+    c_list : list[SubsonicServerConfig] = list()
+    config_index : int
+    for config_index in range(10):
+        config_file_name : str = get_indexed_env_value(ConfigKey.SUBSONIC_PARAMETERS_FILE.getKey(), config_index)
+        server_url : str = get_indexed_env_value(ConfigKey.SUBSONIC_BASE_URL.getKey(), config_index)
+        if config_file_name or server_url:
+            current_config : SubsonicServerConfig = SubsonicServerConfig(config_index)
+            c_list.append(current_config)
+    return c_list
+
